+++ CHANGELOG.md
@@ -1,8 +1,11 @@
 # Changelog
 
+## 1.0.0
+- Support simple mode where authentication is dealt with externally, intended
+  for internal CoreFiling use.
+
 ## 0.1.1
 Correction to base URL of API proxy.
-
 
 ## 0.1.0
 Initial release.